ADPT-108: Review Findings
