+++ taskana-adapter-camunda-listener/pom.xml
@@ -24,6 +24,11 @@
                         <version>2.10.0</version>
                 </dependency>
           <dependency>
+            <groupId>pro.taskana</groupId>
+            <artifactId>taskana-common</artifactId>
+            <version>${version.taskana}</version>
+          </dependency>
+          <dependency>
             <groupId>org.mybatis</groupId>
             <artifactId>mybatis</artifactId>
             <version>${version.mybatis}</version>
