+++ taskana-adapter-camunda-listener/src/main/java/pro/taskana/adapter/camunda/CamundaListenerConfiguration.java
@@ -2,19 +2,18 @@ package pro.taskana.adapter.camunda;
 
 import java.io.FileInputStream;
 import java.io.InputStream;
-import java.time.Duration;
 import java.util.Properties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import pro.taskana.adapter.camunda.outbox.rest.exception.SystemException;
+import pro.taskana.common.api.exceptions.SystemException;
 
-public class OutboxRestConfigurationProperties {
+public class CamundaListenerConfiguration {
 
+  private static final String TASKANA_ADAPTER_CREATE_OUTBOX_SCHEMA =
+      "taskana.adapter.create_outbox_schema";
   private static final String TASKANA_OUTBOX_PROPERTIES = "taskana-outbox.properties";
-
   private static final String TASKANA_ADAPTER_OUTBOX_SCHEMA = "taskana.adapter.outbox.schema";
-
   private static final String TASKANA_ADAPTER_OUTBOX_DATASOURCE_JNDI =
       "taskana.adapter.outbox.datasource.jndi";
   private static final String TASKANA_ADAPTER_OUTBOX_DATASOURCE_DRIVER =
@@ -25,24 +24,19 @@ public class OutboxRestConfigurationProperties {
       "taskana.adapter.outbox.datasource.username";
   private static final String TASKANA_ADAPTER_OUTBOX_DATASOURCE_PASSWORD =
       "taskana.adapter.outbox.datasource.password";
-
-  private static final String TASKANA_ADAPTER_OUTBOX_MAX_NUMBER_OF_EVENTS =
-      "taskana.adapter.outbox.max.number.of.events";
-  private static final String TASKANA_ADAPTER_OUTBOX_DURATION_BETWEEN_TASK_CREATION_RETRIES =
-      "taskana.adapter.outbox.duration.between.task.creation.retries";
-
+  private static final String TASKANA_ADAPTER_OUTBOX_INITIAL_NUMBER_OF_TASK_CREATION_RETRIES =
+      "taskana.adapter.outbox.initial.number.of.task.creation.retries";
   private static final String OUTBOX_SYSTEM_PROPERTY = "taskana.outbox.properties";
-  private static final String OUTBOX_SCHEMA_DEFAULT = "taskana_tables";
-  private static final int MAX_NUMBER_OF_EVENTS_DEFAULT = 50;
-  private static final Duration DURATION_BETWEEN_TASK_CREATION_RETRIES_DEFAULT =
-      Duration.ofHours(1);
 
-  private static final Logger LOGGER =
-      LoggerFactory.getLogger(OutboxRestConfigurationProperties.class);
+  private static final String OUTBOX_SCHEMA_DEFAULT = "taskana_tables";
+  private static final boolean CREATE_OUTBOX_SCHEMA_DEFAULT = true;
+  private static final int INITIAL_NUMBER_OF_TASK_CREATION_RETRIES_DEFAULT = 5;
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(CamundaListenerConfiguration.class);
 
   private final Properties outboxProperties = new Properties();
 
-  private OutboxRestConfigurationProperties() {
+  private CamundaListenerConfiguration() {
 
     String outboxPropertiesFile = System.getProperty(OUTBOX_SYSTEM_PROPERTY);
 
@@ -52,14 +46,14 @@ public class OutboxRestConfigurationProperties {
 
         outboxProperties.load(propertiesStream);
 
-        if (LOGGER.isDebugEnabled()) {
-          LOGGER.debug(
-              String.format("Outbox properties were loaded from file %s.", outboxPropertiesFile));
-        }
+        LOGGER.info(
+            String.format("Outbox properties were loaded from file %s.", outboxPropertiesFile));
+
       } catch (Exception e) {
         LOGGER.warn(
             String.format(
-                "Caught Exception while trying to load properties from provided properties file %s. "
+                "Caught Exception while trying to load properties from "
+                    + "provided properties file %s. "
                     + "Trying to read properties from classpath",
                 outboxPropertiesFile),
             e);
@@ -71,8 +65,22 @@ public class OutboxRestConfigurationProperties {
     }
   }
 
-  public static OutboxRestConfigurationProperties getInstance() {
-    return OutboxRestConfigurationProperties.LazyHolder.INSTANCE;
+  public static CamundaListenerConfiguration getInstance() {
+    return CamundaListenerConfiguration.LazyHolder.INSTANCE;
+  }
+
+  public static boolean getCreateOutboxSchema() {
+
+    String createOutboxSchemaProperty =
+        getInstance().outboxProperties.getProperty(TASKANA_ADAPTER_CREATE_OUTBOX_SCHEMA);
+
+    if ((createOutboxSchemaProperty == null) || createOutboxSchemaProperty.isEmpty()) {
+      LOGGER.info(
+          "Couldn't retrieve boolean property to create schema or not, setting to default ");
+      return CREATE_OUTBOX_SCHEMA_DEFAULT;
+    } else {
+      return !"false".equalsIgnoreCase(createOutboxSchemaProperty);
+    }
   }
 
   public static String getOutboxSchema() {
@@ -108,65 +116,33 @@ public class OutboxRestConfigurationProperties {
     return getInstance().outboxProperties.getProperty(TASKANA_ADAPTER_OUTBOX_DATASOURCE_PASSWORD);
   }
 
-  public static int getOutboxMaxNumberOfEvents() {
+  public static int getInitialNumberOfTaskCreationRetries() {
+    int initialNumberOfTaskCreationRetries;
 
-    int maxNumberOfEventsReturned = 0;
-
-    String maxNumberOfEventsString =
-        getInstance().outboxProperties.getProperty(TASKANA_ADAPTER_OUTBOX_MAX_NUMBER_OF_EVENTS);
-
-    if (maxNumberOfEventsString == null || maxNumberOfEventsString.isEmpty()) {
-      LOGGER.info(
-          "Couldn't retrieve property entry for max number of events to be returned, "
-              + "setting to default ");
-      maxNumberOfEventsReturned = MAX_NUMBER_OF_EVENTS_DEFAULT;
-
-    } else {
-      try {
-        maxNumberOfEventsReturned = Integer.parseInt(maxNumberOfEventsString);
-      } catch (NumberFormatException e) {
-        maxNumberOfEventsReturned = MAX_NUMBER_OF_EVENTS_DEFAULT;
-        LOGGER.warn(
-            String.format(
-                "Attempted to retrieve max number of events to be returned and caught Exception. "
-                    + "Setting default for max number of events to be returned to %d  ",
-                maxNumberOfEventsReturned),
-            e);
-      }
-    }
-    return maxNumberOfEventsReturned;
-  }
-
-  public static Duration getDurationBetweenTaskCreationRetries() {
-
-    String durationBetweentaskCreationRetriesProperty =
+    String initialNumberOfTaskCreationRetriesProperty =
         getInstance()
             .outboxProperties
-            .getProperty(TASKANA_ADAPTER_OUTBOX_DURATION_BETWEEN_TASK_CREATION_RETRIES);
+            .getProperty(TASKANA_ADAPTER_OUTBOX_INITIAL_NUMBER_OF_TASK_CREATION_RETRIES);
 
-    if (durationBetweentaskCreationRetriesProperty == null
-        || durationBetweentaskCreationRetriesProperty.isEmpty()) {
-      LOGGER.info(
-          "Couldn't retrieve property entry for duration between task creation retries, "
-              + "setting to default ");
-      return DURATION_BETWEEN_TASK_CREATION_RETRIES_DEFAULT;
-    } else {
-      try {
-        return Duration.parse(durationBetweentaskCreationRetriesProperty);
-      } catch (Exception e) {
-        LOGGER.warn(
-            String.format(
-                "Attempted to retrieve duration between task creation retries and caught Exception."
-                    + "Setting default to %s ",
-                durationBetweentaskCreationRetriesProperty),
-            e);
-
-        return DURATION_BETWEEN_TASK_CREATION_RETRIES_DEFAULT;
-      }
+    try {
+      initialNumberOfTaskCreationRetries =
+          Integer.parseInt(initialNumberOfTaskCreationRetriesProperty);
+    } catch (NumberFormatException e) {
+      initialNumberOfTaskCreationRetries = INITIAL_NUMBER_OF_TASK_CREATION_RETRIES_DEFAULT;
+      LOGGER.warn(
+          String.format(
+              "Attempted to retrieve initial number of task creation retries and caught "
+                  + "Exception. Setting default for initial number of "
+                  + "task creation retries to %d ",
+              initialNumberOfTaskCreationRetries),
+          e);
     }
+
+    return initialNumberOfTaskCreationRetries;
   }
 
   private void readPropertiesFromClasspath() {
+
     try (InputStream propertiesStream =
         this.getClass().getClassLoader().getResourceAsStream(TASKANA_OUTBOX_PROPERTIES)) {
 
@@ -193,7 +169,6 @@ public class OutboxRestConfigurationProperties {
   }
 
   private static class LazyHolder {
-    private static final OutboxRestConfigurationProperties INSTANCE =
-        new OutboxRestConfigurationProperties();
+    private static final CamundaListenerConfiguration INSTANCE = new CamundaListenerConfiguration();
   }
 }
