--- taskana-adapter-camunda-listener/src/main/java/pro/taskana/adapter/camunda/exceptions/SystemException.java
@@ -1,15 +1,0 @@
-package pro.taskana.adapter.camunda.exceptions;
-
-/** This exception is thrown when a generic problem is encountered. */
-public class SystemException extends RuntimeException {
-
-  private static final long serialVersionUID = 1L;
-
-  public SystemException(String msg) {
-    super(msg);
-  }
-
-  public SystemException(String msg, Throwable cause) {
-    super(msg, cause);
-  }
-}
